refactor and add dev command
