+++ libs/wingc/src/lsp/completions.rs
@@ -1,6 +1,10 @@
-use lsp_types::{CompletionItemKind, Position};
+use lsp_types::{CompletionItem, CompletionItemKind, CompletionResponse, Position};
 use tree_sitter::{Node, Point, Tree};
 use tree_sitter_traversal::{traverse, Order};
+
+use crate::wasm_util::{ptr_to_string, string_to_combined_ptr};
+
+use crate::lsp::sync::FILES;
 
 #[derive(Debug)]
 pub struct WingCompletionItem {
@@ -84,4 +88,45 @@ pub fn completions_from_ast(source: &str, tree: &Tree, position: Position) -> Ve
 	}
 
 	return completions;
+}
+
+#[no_mangle]
+pub unsafe extern "C" fn wingc_on_completion(ptr: u32, len: u32) -> u64 {
+	let parse_string = ptr_to_string(ptr, len);
+	if let Ok(parsed) = serde_json::from_str(&parse_string) {
+		let result = on_completion(parsed);
+		let result = serde_json::to_string(&result).unwrap();
+
+		// return result as u64 with ptr and len
+		string_to_combined_ptr(result)
+	} else {
+		panic!("Failed to parse 'completion': {}", parse_string);
+	}
+}
+pub fn on_completion(params: lsp_types::CompletionParams) -> CompletionResponse {
+	FILES.with(|files| {
+		let files = files.borrow();
+		let files = files.read().unwrap();
+		let uri = params.text_document_position.text_document.uri;
+		let result = files.get(&uri).unwrap();
+
+		let position = params.text_document_position.position;
+		let completions = completions_from_ast(&result.contents.as_str(), &result.tree, position);
+
+		CompletionResponse::Array(
+			completions
+				.iter()
+				.map(|item| {
+					let text = item.text.as_str();
+					CompletionItem {
+						label: text.to_string(),
+						insert_text: Some(text.to_string()),
+						kind: Some(item.kind),
+						detail: item.detail.clone(),
+						..Default::default()
+					}
+				})
+				.collect(),
+		)
+	})
 }
